# expression: implement vectorized evaluation for builtinNowWithArgSig