@qw4990 2019-11-12
Hi, @ekalinin , we noticed that recently you are very active in our community and have contributed 15 PRs for the expression package!!! Thanks a lot for your contribution and we'd like to invite you to join our [`sig-expr Slack channel`](https://tidbcommunity.slack.com/archives/CMRD79DRR) so that if you have any questions related to expression, you can discuss with us and other contributors. Have a nice day!